integrated different order types
so in the future we can have online and manual orders (e.g. via phone).
as well as free tickets which will be collected under a separate order
type. furthermore there will be statistics for retail tickets. they are
not treated as orders but will have an own order type in statistics.